fix: yarn v1 npm commands call fix
Closes #58
