+++ packages/pvm-core/lib/node-boot.ts
@@ -2,6 +2,14 @@ import { error, logger } from './logger'
 import shell from './shell'
 import { env } from './env'
 import { gitFetch } from './git/commands'
+import path from 'path'
+import fs from 'fs'
+import ini from 'ini'
+
+function readNpmRc(cwd: string): string {
+  const npmrcPath = path.join(cwd, '.npmrc')
+  return fs.existsSync(npmrcPath) ? fs.readFileSync(npmrcPath).toString('utf8') : ''
+}
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { version } = require('../package.json')
@@ -54,6 +62,21 @@ function nodeBoot(): void {
       shell('git rev-parse origin/master', { cwd })
     } catch (e) {
       gitFetch(cwd)
+    }
+  }
+
+  const npmRc = readNpmRc(cwd)
+  // eslint-disable-next-line pvm/no-process-env
+  const classicYarn = process.env.npm_config_user_agent ? process.env.npm_config_user_agent.indexOf('yarn/1') !== -1 : false
+  // yarn 1 overrides npmrc settings with its own set of npm config envs so `yarn publish-command` wont work if it depends on those
+  // strict-ssl is one of those important settings, so copy it to envs and override yarn default one
+  if (classicYarn && npmRc) {
+    if (npmRc.indexOf('strict-ssl') !== -1) {
+      const parsedNpmRc = ini.parse(npmRc)
+      if (parsedNpmRc['strict-ssl'] !== undefined) {
+        // eslint-disable-next-line pvm/no-process-env
+        process.env.npm_config_strict_ssl = parsedNpmRc['strict-ssl']
+      }
     }
   }
 }
