+++ src/plugins/core/package.json
@@ -22,7 +22,6 @@
     "chalk": "^4.0.0",
     "cli-table": "^0.3.6",
     "p-map": "^4.0.0",
-    "lodash": "^4.2.0",
-    "ini": "^1.3.8"
+    "lodash": "^4.2.0"
   }
 }
