+++ test/scenarios/__fixtures__/boot-test.js
@@ -1,0 +1,2 @@
+process.chdir(process.argv[1])
+require('@pvm/core/lib/node-boot')
